fix #1, fix #2

+++ src/main/java/dev/mrsterner/guardvillagers/common/entity/ai/goals/WalkBackToCheckPointGoal.java
@@ -31,7 +31,7 @@ public class WalkBackToCheckPointGoal extends Goal {
         BlockPos blockpos = this.guard.getPatrolPos();
         if (blockpos != null) {
             Vec3d vector3d = Vec3d.ofBottomCenter(blockpos);
-            Vec3d vector3d1 = NoPenaltyTargeting.findTo(this.guard, 16, 3, vector3d, (double) ((float) Math.PI / 10F));
+            Vec3d vector3d1 = NoPenaltyTargeting.findTo(this.guard, 16, 3, vector3d, ((float) Math.PI / 10F));
             if (!this.guard.getPatrolPos().isWithinDistance(this.guard.getPos(), 1.0D) || blockpos != null)
                 if (guard.getMainHandStack().getItem() instanceof RangedWeaponItem) {
                     this.guard.getNavigation().startMovingTo(blockpos.getX(), blockpos.getY(), blockpos.getZ(), this.speed);
